Rename database.ts into provinces.ts
